+++ demo/src/main.tsx
@@ -9,6 +9,7 @@ import PopupView from "./components/popup/popupView";
 import iconGit from "./assets/gitIcon.svg";
 import imgLogo from "./assets/logo.png";
 import styles from "./main.scss";
+
 import ControlsView from "./components/controls/controlsView";
 import SpinView from "./components/spin/spinView";
 import TextControlView from "./components/controls/text";
@@ -20,8 +21,9 @@ import CheckControlView from "./components/controls/check";
 import RadioControlView from "./components/controls/radio";
 import CalendarControlView from "./components/controls/calendar";
 import DateControlView from "./components/controls/date";
-import FAQView from "./components/FAQView";
+import TimeControlView from "./components/controls/time";
 import NumberControlView from "./components/controls/number";
+import FAQView from "./components/FAQView";
 
 (window as any).WUPHelpers = WUPHelpers;
 
@@ -49,6 +51,7 @@ const routes: IRoute[] = [
   // { path: "control/selectMany", el: SelectManyControlView, isNested: true },
   { path: "control/calendar", el: CalendarControlView, isNested: true },
   { path: "control/date", el: DateControlView, isNested: true },
+  { path: "control/time", el: TimeControlView, isNested: true },
   { path: "faq", label: "FAQ", el: FAQView },
 ];