+++ src/index.ts
@@ -3,6 +3,7 @@ import * as WUPHelpers from "./indexHelpers";
 export { WUPHelpers };
 
 export { default as WUPTimeObject } from "./objects/timeObject";
+export { default as localeInfo } from "./objects/localeInfo";
 
 // export { default as WUPBaseElement } from "./baseElement";
 export { default as WUPPopupElement } from "./popup/popupElement";
@@ -18,4 +19,5 @@ export { default as WUPCheckControl } from "./controls/check";
 export { default as WUPRadioControl } from "./controls/radio";
 export { default as WUPCalendarControl } from "./controls/calendar";
 export { default as WUPDateControl } from "./controls/date";
+export { default as WUPTimeControl } from "./controls/time";
 export { default as WUPNumberControl } from "./controls/number";